+++ grails-app/services/io/xh/hoist/monitor/MonitoringService.groovy
@@ -12,10 +12,10 @@ import grails.events.EventPublisher
 import grails.gorm.transactions.ReadOnly
 import io.xh.hoist.BaseService
 import io.xh.hoist.util.Timer
-import io.xh.hoist.async.AsyncSupport
 
 import java.util.concurrent.ConcurrentHashMap
 
+import static grails.async.Promises.task
 import static io.xh.hoist.monitor.MonitorStatus.*
 import static io.xh.hoist.util.DateTimeUtils.MINUTES
 import static io.xh.hoist.util.DateTimeUtils.SECONDS
@@ -32,7 +32,7 @@ import static java.lang.System.currentTimeMillis
  *
  * If enabled via config, this service will also write monitor run results to a dedicated log file.
  */
-class MonitoringService extends BaseService implements AsyncSupport, EventPublisher {
+class MonitoringService extends BaseService implements EventPublisher {
 
     def configService,
         monitorResultService
@@ -81,7 +81,7 @@ class MonitoringService extends BaseService implements AsyncSupport, EventPublis
             def timeout = getTimeoutSeconds()
 
             def tasks = Monitor.list().collect { m ->
-                asyncTask { monitorResultService.runMonitor(m, timeout) }
+                task { monitorResultService.runMonitor(m, timeout) }
             }
 
             Map newResults = Promises